refactor: extract validation and rendering helpers from engine/prompt.py
