+++ tests/unit/engine/test_run_result.py
@@ -16,7 +16,8 @@ from synthorg.engine.loop_protocol import (
     TurnRecord,
     make_budget_checker,
 )
-from synthorg.engine.prompt import SystemPrompt, format_task_instruction
+from synthorg.engine.prompt import SystemPrompt
+from synthorg.engine.prompt_validation import format_task_instruction
 from synthorg.engine.run_result import AgentRunResult
 from synthorg.providers.enums import FinishReason, MessageRole
 from synthorg.providers.models import ChatMessage, TokenUsage, ToolCall
